requirements.txt
faiss-cpu
sentence-transformers
pdfplumber
pytesseract
pillow
pandas
numpy
streamlit
